# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Glasshaus API",
    "description": "REST API for Project Glasshaus. Every capability is also available as an MCP tool. Authenticate with `Authorization: Bearer <api token>` or the browser session cookies (unsafe methods then need the `X-CSRF-Token` header). Errors use RFC 9457 problem details.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List projects you can see",
        "operationId": "list_projects",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectRead"
                  },
                  "title": "Response List Projects"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create a project",
        "operationId": "create_project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/projects/by-key/{key}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get a project by its key",
        "operationId": "get_project_by_key",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get a project with its statuses",
        "operationId": "get_project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update or archive a project",
        "operationId": "update_project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete a project and its tasks (dry_run=true, the default, only previews)",
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Dry Run"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePreview"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/members": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List project members",
        "operationId": "list_project_members",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectMemberRead"
                  },
                  "title": "Response List Project Members"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Add or change a project member",
        "operationId": "set_project_member",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectMemberSet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/members/{user_id}": {
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Remove a project member",
        "operationId": "remove_project_member",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/statuses": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List workflow statuses",
        "operationId": "list_statuses",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusRead"
                  },
                  "title": "Response List Statuses"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add a status",
        "operationId": "create_status",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem"
      },
      "ProjectCreate": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "key": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9]{1,9}$",
            "title": "Key",
            "description": "Short uppercase key used in task references, e.g. WEB-12."
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "title": "Description",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workspace_id",
          "key",
          "name"
        ],
        "title": "ProjectCreate"
      },
      "ProjectDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "my_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRole"
              },
              {
                "type": "null"
              }
            ],
            "description": "The caller's effective role on this project."
          },
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/StatusRead"
            },
            "type": "array",
            "title": "Statuses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "key",
          "name",
          "description",
          "archived_at",
          "created_at",
          "updated_at",
          "statuses"
        ],
        "title": "ProjectDetail"
      },
      "ProjectMemberRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/ProjectRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "role"
        ],
        "title": "ProjectMemberRead"
      },
...
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "my_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRole"
              },
              {
                "type": "null"
              }
            ],
            "description": "The caller's effective role on this project."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "key",
          "name",
          "description",
          "archived_at",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectRead"
      },
      "ProjectRole": {
        "type": "string",
        "enum": [
          "admin",
          "editor",
          "commenter",
          "viewer"
        ],
        "title": "ProjectRole"
      },
...
    }
  }
}